README.md
# Project 1
A web homepage for "Aloha Apparel Co." using HTML, CSS, and JS.

This was a large project that incorporated all the web development skills I learned over my first weeks in this career.

The Aloha webpage is built with HTML, CSS with Javascript for interactivity.
Design comparables of 3 different screen sizes were provided for reference.
The final product is a mobile first-styled responsive web design, though the project was actually done with the largest, desktop, resolution completed first. It was certainly an extra challenge making the design mobile-first belated, a mistake I'll be keen to avoid in the future.

The webpage uses HTML5 semantic markup (head, header, section, footer, etc.) and incorporates a list-based navigation menu. Initially, I was using too many div containers in place of section elements, and  IDs were not used as selectors (over classes) until the necessity for internal navigation, but I see the advantages of using a variety now.

The webpage largely uses flex box layout properties because it is simply put, the easier way to program (and a project constraint on not using floats but even with it, flexbox all the way). I got to put hours of Flexbox Zombies practice into use towards every section of the webpage, most notably in the Shopping menu grid that utilizes flexwrap.

Custom fonts are incorporated by referencing a separate font stylesheet, prepared by fontsquirrel.com

A basic javascript code provides alert responses to the email/newsletter form submission, and with plugins, adds a "smoothscroll" for the list navigation menu, and a carousel product browser with "Flickity".